// src/modules/textile/reports/knitting/knitting-program-report/knitting-program-report-type.ts
export type IKnittingProgramReport = {
  KNITTING_PROGRAM_NO: string;
  KNITTING_PROG_DATE: Date;
  COMPANY_ID: number;
  COMPANY_NAME: string;
  COMPANY_ADDRESS: string;
  KNITTING_HOUSE_ID: number;
  KNIT_HOUSE: string;
  KNIT_HOUSE_ADDRESS: string;
  BUYERID: number;
  BUYER: string;
  PO_ID: number;
  PONO: string;
  STYLE_ID: number;
  STYLENO: string;
  YARN: string;
  YARN_LOT_NUMBER: string;
  YARN_COUNT: string;
  FABRIC: string;
  FABRIC_TYPE: string;
  MC_DIA: string;
  FINISH_DIA: string;
  GSM: string;
  GMT_COLOR_ID: number;
  COLORNAME: string;
  YARN_COLOR: string;
  STITCH_LENGTH: string;
  FABRIC_ID: number;
  GAUGE: string;
  KNITTING_RATE: number;
  START_DATE: Date;
  END_DATE: Date;
  REMARKS: string;
  LYCRA_CM: string;
  QUANTITY: number;
  PICES: number;
  TOTAL_AMOUNT: number;
  KNITTING_ADVICE: string;
  PREPARED_BY: string;
  FABRIC_PART: string;
  STRIPE_MEASUREMENT_CM: string;
  CPI: string;
  PLY: string;
  NEEDLE_RATIO: string;
};